Rename ext/location.* to ext/legacy_location.* to avoid name conflicts
